Game WIP
Added 'win' to classes removed when game is reset.
Added function to mark the winning combination on the gameboard.
Refactored _checkWinner to use a traditional for loop so that we can break out of it if a winner in found, since underscore's .each() implementation doesn't allow the use of break.
Added call to _markWinningCombinatino if a winner is declared.